requirements.txt
audioop-lts==0.2.1
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
comtypes==1.4.10
docopt==0.6.2
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.70.0
greenlet==3.2.0
httplib2==0.22.0
idna==3.10
Js2Py==0.74
mysql-connector-python==9.3.0
oauthlib==3.2.2
packaging==24.2
pipwin==0.5.2
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pydub==0.25.1
pygame==2.6.1
pyjsparser==2.7.1
pyparsing==3.2.3
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
python-vlc==3.0.21203
pyttsx3==2.98
pytube==15.0.0
pywin32==310
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
soupsieve==2.6
SpeechRecognition==3.14.2
SQLAlchemy==2.0.40
standard-aifc==3.13.0
standard-chunk==3.13.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
wikipedia==1.4.0
yt-dlp==2025.3.31
audioop-lts==0.2.1
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
comtypes==1.4.10
docopt==0.6.2
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.70.0
greenlet==3.2.0
httplib2==0.22.0
idna==3.10
Js2Py==0.74
mysql-connector-python==9.3.0
oauthlib==3.2.2
packaging==24.2
pipwin==0.5.2
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pydub==0.25.1
pygame==2.6.1
pyjsparser==2.7.1
pyparsing==3.2.3
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
python-vlc==3.0.21203
pyttsx3==2.98
pytube==15.0.0
pywin32==310
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
soupsieve==2.6
SpeechRecognition==3.14.2
SQLAlchemy==2.0.40
standard-aifc==3.13.0
standard-chunk==3.13.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
wikipedia==1.4.0
yt-dlp==2025.3.31
audioop-lts==0.2.1
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
comtypes==1.4.10
diskcache==5.6.3
docopt==0.6.2
filelock==3.18.0
fsspec==2025.3.2
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.70.0
greenlet==3.2.0
httplib2==0.22.0
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
Js2Py==0.74
llama_cpp_python==0.3.8
MarkupSafe==3.0.2
mpmath==1.3.0
mysql-connector-python==9.3.0
networkx==3.4.2
numpy==2.2.5
oauthlib==3.2.2
packaging==24.2
pipwin==0.5.2
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pydub==0.25.1
pygame==2.6.1
pyjsparser==2.7.1
pyparsing==3.2.3
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
python-vlc==3.0.21203
pyttsx3==2.98
pytube==15.0.0
pywin32==310
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.5.3
setuptools==80.3.0
six==1.17.0
soupsieve==2.6
SpeechRecognition==3.14.2
SQLAlchemy==2.0.40
standard-aifc==3.13.0
standard-chunk==3.13.0
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
wikipedia==1.4.0
yt-dlp==2025.3.31
